Extracted test fixtures
