+++ tests/CommandBus/CommandBusTest.php
@@ -8,8 +8,8 @@ use App\Commands\MyCommand;
 use function Tempest\command;
 use Tempest\Commands\CommandBus;
 use Tempest\Commands\CommandBusConfig;
-use Tempest\Commands\CommandBusMiddleware;
 use Tempest\Commands\CommandHandlerNotFound;
+use Tests\Tempest\CommandBus\Fixtures\MyCommandBusMiddleware;
 use Tests\Tempest\TestCase;
 
 class CommandBusTest extends TestCase
@@ -46,17 +46,5 @@ class CommandBusTest extends TestCase
         $this->expectException(CommandHandlerNotFound::class);
 
         command(new class () {});
-    }
-}
-
-class MyCommandBusMiddleware implements CommandBusMiddleware
-{
-    public static bool $hit = false;
-
-    public function __invoke(object $command, callable $next): void
-    {
-        self::$hit = true;
-
-        $next($command);
     }
 }
