+++ tests/CommandBus/Fixtures/MyCommandBusMiddleware.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\CommandBus\Fixtures;
+
+use Tempest\Commands\CommandBusMiddleware;
+
+class MyCommandBusMiddleware implements CommandBusMiddleware
+{
+    public static bool $hit = false;
+
+    public function __invoke(object $command, callable $next): void
+    {
+        self::$hit = true;
+
+        $next($command);
+    }
+}
