+++ tests/Console/RenderConsoleCommandTest.php
@@ -10,6 +10,7 @@ use function Tempest\attribute;
 use Tempest\Console\ConsoleCommand;
 use Tempest\Console\ConsoleStyle;
 use Tempest\Console\RenderConsoleCommand;
+use Tests\Tempest\Console\Fixtures\MyConsole;
 
 class RenderConsoleCommandTest extends TestCase
 {
@@ -37,19 +38,5 @@ class RenderConsoleCommandTest extends TestCase
             'test <path> [times=1] [--force=false] - description',
             $string
         );
-    }
-}
-
-class MyConsole
-{
-    #[ConsoleCommand(
-        name: 'test',
-        description: 'description',
-    )]
-    public function handle(
-        string $path,
-        int $times = 1,
-        bool $force = false,
-    ) {
     }
 }
