+++ tests/Container/Fixtures/BuiltinArrayClass.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Container\Fixtures;
+
+class BuiltinArrayClass
+{
+    public function __construct(public array $anArray)
+    {
+    }
+}
