+++ tests/Container/Fixtures/CallContainerObjectE.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Container\Fixtures;
+
+use Tests\Tempest\Container\Fixtures;
+
+class CallContainerObjectE
+{
+    public function method(Fixtures\ContainerObjectE $input): Fixtures\ContainerObjectE
+    {
+        return $input;
+    }
+}
