+++ tests/Container/Fixtures/ContainerObjectA.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Container\Fixtures;
+
+class ContainerObjectA
+{
+}
