+++ tests/Container/Fixtures/ContainerObjectC.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Container\Fixtures;
+
+class ContainerObjectC
+{
+    public function __construct(public string $prop)
+    {
+    }
+}
