+++ tests/Container/Fixtures/ContainerObjectE.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Container\Fixtures;
+
+class ContainerObjectE
+{
+    public function __construct(public string $id = 'default')
+    {
+    }
+}
