+++ tests/Container/Fixtures/SingletonClass.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Container\Fixtures;
+
+class SingletonClass
+{
+    public static int $count = 0;
+
+    public function __construct()
+    {
+        self::$count += 1;
+    }
+}
