+++ tests/Container/Fixtures/UnionTypesClass.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Container\Fixtures;
+
+use DateTime;
+
+class UnionTypesClass
+{
+    public function __construct(
+        public DateTime $aStringOrDate
+    ) {
+    }
+}
