+++ tests/Events/EventBusTest.php
@@ -9,7 +9,7 @@ use App\Events\MyEventHandler;
 use function Tempest\event;
 use Tempest\Events\EventBus;
 use Tempest\Events\EventBusConfig;
-use Tempest\Events\EventBusMiddleware;
+use Tests\Tempest\Events\Fixtures\MyEventBusMiddleware;
 use Tests\Tempest\TestCase;
 
 class EventBusTest extends TestCase
@@ -38,18 +38,5 @@ class EventBusTest extends TestCase
         event(new ItHappened());
 
         $this->assertTrue(MyEventBusMiddleware::$hit);
-    }
-}
-
-
-class MyEventBusMiddleware implements EventBusMiddleware
-{
-    public static bool $hit = false;
-
-    public function __invoke(object $event, callable $next): void
-    {
-        self::$hit = true;
-
-        $next($event);
     }
 }
