+++ tests/Mapper/Fixtures/ObjectFactoryA.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\Mapper\Fixtures;
+
+use Tempest\ORM\Attributes\CastWith;
+use Tempest\ORM\IsModel;
+use Tempest\ORM\Model;
+
+class ObjectFactoryA implements Model
+{
+    use IsModel;
+
+    #[CastWith(ObjectFactoryACaster::class)]
+    public string $prop;
+}
