+++ tests/ORM/Fixtures/Foo.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Tempest\ORM\Fixtures;
+
+use Tempest\ORM\IsModel;
+use Tempest\ORM\Model;
+
+class Foo implements Model
+{
+    use IsModel;
+
+    public string $bar;
+}
