+++ tests/Validation/ValidatorTest.php
@@ -6,9 +6,8 @@ namespace Tests\Tempest\Validation;
 
 use PHPUnit\Framework\TestCase;
 use Tempest\Validation\Exceptions\ValidationException;
-use Tempest\Validation\IsValidated;
-use Tempest\Validation\Rules\Length;
 use Tempest\Validation\Validator;
+use Tests\Tempest\Validation\Fixtures\ObjectTobeValidated;
 
 class ValidatorTest extends TestCase
 {
@@ -20,14 +19,5 @@ class ValidatorTest extends TestCase
         $validator = new Validator();
 
         $validator->validate(new ObjectTobeValidated(name: 'a'));
-    }
-}
-
-class ObjectTobeValidated implements IsValidated
-{
-    public function __construct(
-        #[Length(min: 2, max: 3)]
-        public string $name,
-    ) {
     }
 }
